flaskdbexemple/requirements.txt
alembic==1.4.2
attrs==19.3.0
blinker==1.4
click==7.1.2
coverage==5.1
DateTime==4.3
decorator==4.4.2
fields==5.0.0
Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
flask-oidc==1.4.0
Flask-RESTful==0.3.8
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
Flask-Testing==0.8.0
httplib2==0.18.1
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
linecache2==1.0.0
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.6.1
marshmallow-jsonapi==0.23.1
marshmallow-sqlalchemy==0.23.1
nose2==0.9.2
oauth2client==4.1.3
pbr==5.4.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
PyMySQL==0.9.3
pyrsistent==0.16.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
resources==0.0.1
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.18
sqlalchemy-migrate==0.13.0
sqlparse==0.3.1
Tempita==0.5.2
traceback2==1.4.0
unittest2==1.1.0
uuid==1.30
Werkzeug==1.0.1
zipp==3.1.0
zope.interface==5.1.0